requirements.txt
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
pypdf
python-docx
PyYAML
whoosh
fastapi
uvicorn[standard]
httpx
